+++ scripts/pipeline/seed-data.sh
@@ -13,10 +13,10 @@ function main() {
 
   aws s3 cp s3://participating-icb/Participating_ICBs.csv ./test-data
 
-
   source $PWD/scripts/pipeline/create-data-files.sh
 
   echo Finished creating data
+
   aws dynamodb batch-write-item --request-items \
           file://$PWD/test-data/participating_icb.json
 }
